requirements.txt
scikit-learn==1.3.0
snowflake-connector-python==3.0.4
snowflake-snowpark-python[pandas]==1.5.1
python-dotenv==1.0.0
pandas==1.4.1
PyYAML==6.0
plotly-express==0.4.1
kprototypes==0.1.1
gower==0.1.2
kneed==0.8.5
umap-learn==0.5.3
matplotlib==3.7.2